requirements.txt
django>=3.0.2,<3.0.99
djangorestframework>=3.11.0,<3.11.99
psycopg2-binary>=2.8.4,<2.8.99
Pillow>=7.0.0,<7.0.99
whitenoise>=5.0.1,<5.0.99
drf-yasg>=1.17.0,<1.17.99
factory-boy>=2.12.0,<2.12.99
dj-inmemorystorage>=2.0.0,<2.0.99
flake8>=3.7.9,<3.7.99
isort>=4.3.21,<4.3.99
black==19.10b0
celery>=4.4.0,<4.4.99
coverage>=5.0.3,<5.0.99
django-otp>=0.7.5,<0.7.99
qrcode==6.1
django-redis>=4.5.0,<4.5.99